fix: chatRoomList 초기화

+++ Backend/IdolBom/src/main/java/Fav_I/IdolBom/Entity/ParticipantListId.java
@@ -16,7 +16,7 @@ public class ParticipantListId implements java.io.Serializable {
     private static final long serialVersionUID = 8625574425690214356L;
     @NotNull
     @Column(name = "userID", nullable = false)
-    private Integer userID;
+    private Long userID;
 
     @NotNull
     @Column(name = "recruitmentID", nullable = false)
